Divide post modal for post editor & creating post, fix some UI bugs with option button
