+++ src/conponents/account-setting/AccountSettingNavigation.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { NavLink } from 'react-router-dom'
-import styled from 'styled-components'
 
 function AccountSettingNavigation() {
     return (
